Sync

+++ util/math.go
@@ -14,7 +14,7 @@ import (
 // Helpers
 
 // valToPercent converts 1.05 to 5 or 0.95 to 95
-func valToPercent(v int) int {
+func valToPercent(v float64) float64 {
 	if v > 1 {
 		return (v - 1) * 100
 	} else {
